Move ArrayWithOffset class to ZIL module
