+++ samples/99_zil_interpreter/app/builtins.rb
@@ -339,7 +339,7 @@ ZIL_BUILTINS[:NTH] = define_for_evaled_arguments { |arguments|
 ZIL_BUILTINS[:REST] = define_for_evaled_arguments { |arguments|
   array_like_value = arguments[0]
   offset = arguments[1] || 1
-  ArrayWithOffset.from(array_like_value, offset: offset)
+  ZIL::ArrayWithOffset.from(array_like_value, offset: offset)
 }
 
 ZIL_BUILTINS[:BACK] = lambda { |arguments, context|
@@ -370,48 +370,3 @@ ZIL_BUILTINS[:PUTREST] = define_for_evaled_arguments { |arguments|
   array_like_value[1..-1] = new_rest.dup
   array_like_value
 }
-
-class ArrayWithOffset
-  attr_reader :original_array, :offset
-
-  def initialize(original_array, offset:)
-    @original_array = original_array
-    @offset = offset
-  end
-
-  def [](index)
-    @original_array[with_offset(index)]
-  end
-
-  def []=(index, value)
-    @original_array[with_offset(index)] = value
-  end
-
-  def to_a
-    @original_array[@offset..-1]
-  end
-
-  def self.from(value, offset:)
-    case value
-    when Array
-      new(value, offset: offset)
-    when ArrayWithOffset
-      new(value.original_array, offset: offset + value.offset)
-    else
-      raise "REST not supported for #{value}"
-    end
-  end
-
-  private
-
-  def with_offset(index)
-    case index
-    when Integer
-      return index if index.negative? # Don't offset index relative to the end
-
-      index + @offset
-    when Range
-      Range.new(with_offset(index.begin), with_offset(index.end), index.exclude_end?)
-    end
-  end
-end
