+++ samples/99_zil_interpreter/app/main.rb
@@ -1,4 +1,4 @@
-require 'zil/table.rb'
+require 'zil.rb'
 require 'app/syntax.rb'
 require 'app/parser.rb'
 require 'app/builtins.rb'
