+++ samples/99_zil_interpreter/zil.rb
@@ -1,0 +1,2 @@
+require 'zil/array_with_offset.rb'
+require 'zil/table.rb'
